feat: add interface-stream-muxer support and tests
