+++ src/codec.js
@@ -1,0 +1,3 @@
+'use strict'
+
+module.exports = '/mplex/6.7.0'
