+++ test/compliance.spec.js
@@ -1,0 +1,16 @@
+/* eslint-env mocha */
+'use strict'
+
+const tests = require('interface-stream-muxer')
+const multiplex = require('../src')
+
+describe('compliance', () => {
+  tests({
+    setup (cb) {
+      cb(null, multiplex)
+    },
+    teardown (cb) {
+      cb()
+    }
+  })
+})
